fix: switch between authenticated and unauthenticated modes
