+++ src/styles/App.css
@@ -81,7 +81,7 @@ input.stay-days {
 .right-column {
   overflow-y: hidden;
   height: 100vh;
-  padding: 5rem 1rem 0;
+  padding: 4rem 1rem 0;
   min-width: 14rem;
 }
