+++ src/views/MapView/components/CountryPopup.jsx
@@ -7,7 +7,6 @@ import {
   CompassOutlined,
   PlusCircleOutlined,
   LikeOutlined,
-  CheckOutlined,
   FolderOutlined
 } from "@ant-design/icons";
 import { useAuthContext } from "../../../context/AuthContext";
@@ -15,9 +14,9 @@ import { useAppModal } from "../../../components/AppModal";
 import CreateNewVisit from "../../../components/Modals/CreateNewVisit";
 import useLoadHistory from "../../../api/history/useLoadHistory";
 import useLoadMeWithGroups from "../../../api/useLoadMeWithGroups";
-import useUpdateMeWithGroups from "../../../api/useUpdateMeWithGroups";
-import { message } from "antd";
 import shortMonthToLongMonth from "../../../helpers/shortMonthToLongMonth";
+import AddGroups from "../../../components/Modals/AddGroup";
+
 const PopupGroup = ({ name }) => {
   return (
     <div
@@ -30,118 +29,17 @@ const PopupGroup = ({ name }) => {
   )
 }
 
-export const AddGroups = ({ country }) => {
-  const { data, loading, error } = useLoadMeWithGroups();
-  const { data: dataPut, executePutGroups, loading: putLoading, error: putError } = useUpdateMeWithGroups();
-  const [dataWithProperties, setDataWithProperties] = useState();
-  const modal = useAppModal();
-  useEffect(() => {
-
-    !loading
-      && setDataWithProperties(data.groups.map((group) => ({ ...group, isSelected: group.regions.some((country1) => country1.id === country.id) })))
-  }, [loading])
-  return (
-    <div>
-      <h4 className={'fs-5 fw-bold'}>Add Groups</h4>
-      {dataWithProperties && !loading && <div>
-        <div style={{ marginTop: "2rem", display: "grid", gap: "1rem", gridTemplateColumns: "0.25fr 0.25fr 0.25fr 0.25fr" }}>
-          {dataWithProperties.map((group, index) => {
-            return (
-              <div
-                key={index}
-                onClick={() => {
-                  setDataWithProperties(
-                    dataWithProperties.map(
-                      (group1) => group1.id !== group.id ? group1 :
-                        { ...group, isSelected: !group.isSelected }
-                    )
-                  )
-                }}
-                style={{ cursor: "pointer", border: group.isSelected ? "1px solid #FFFFFF" : "1px solid #336273", display: "grid", width: "162px", height: "32px", backgroundImage: "linear-gradient(to right,#0B1C22 0%,#1B404D 100%)", borderRadius: "18px" }}>
-                <div style={{ paddingLeft: "1rem", paddingRight: "1rem", display: "flex", alignItems: "center", justifyContent: "space-between" }}>
-                  <span style={{ fontSize: "10px", fontWeight: "Bold" }}>
-                    {group.name}
-                  </span>
-                  <div style={{ display: "flex", justifyContent: "center", alignItems: "center", height: "1.3rem", border: group.isSelected ? "1px solid #FFFFFF" : "1px solid transparent", width: "1.3rem", backgroundColor: "#336273", borderRadius: "50%" }}>
-                    <CheckOutlined style={{ display: group.isSelected ? "block" : "none", fontSize: "10px" }} />
-                  </div>
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      </div>
-      }
-      <div style={{ marginTop: "2rem", justifyContent: "flex-end", display: "flex" }}>
-        <span style={{ marginRight: "1rem", alignContent: "center", cursor: "pointer" }}
-          onClick={() => {
-            modal.reset()
-          }}
-        >
-          Cancel
-        </span>
-        <div
-          onClick={() => {
-            setDataWithProperties(dataWithProperties
-              .map((group) => ({
-                name: group.name,
-                regions: group.isSelected
-                  ?
-                  [{ id: country.id }, ...group.regions.map((country1) => ({ id: country1.id }))]
-                  :
-                  [...group.regions.map((country1) => ({ id: country1.id }))]
-              })))
-
-            executePutGroups(
-              {
-                data: {
-                  "groups": dataWithProperties
-                    .map((group) =>
-                    ({
-                      name: group.name,
-                      regions: group.isSelected
-                        ?
-                        [{ id: country.id }, ...group.regions
-                          .filter((filteredCountry) => filteredCountry.id !== country.id)
-                          .map((country1) => ({ id: country1.id }))]
-                        :
-                        [...group.regions
-                          .filter((filteredCountry) => filteredCountry.id !== country.id)
-                          .map((country1) => ({ id: country1.id }))]
-                    }))
-                }
-              }).then((e) => {
-                message.success("Groups updated successfully");
-                modal.reset()
-              })
-              .catch((error) => {
-                message.error("Groups update failed")
-              })
-          }}
-          style={{ cursor: "pointer", display: "flex", justifyContent: "center", alignItems: "center", position: "relative", color: "black", height: "2.3rem", minWidth: "11.5rem", background: "white", borderRadius: "0.8rem", marginRight: "1rem" }}>
-          Add groups
-        </div>
-      </div>
-    </div>
-  );
-
-}
-
 
 export const CountryPopup = ({ country }) => {
-  const { data, fetch } = useLoadMeWithGroups();
   const [selectedGroup, setSelectedGroup] = useState();
   const iconStyle = { fontSize: '12px' };
   const { user } = useAuthContext();
+  const { data, fetch } = useLoadMeWithGroups();
   const modal = useAppModal();
   const { data: historyData, getDataForTheRegion: getHistoryDataForRegion } = useLoadHistory({
     userId: user?.id,
     regionId: country.country.id,
   });
-
-  useEffect(() => {
-    getHistoryDataForRegion(country.id);
-  }, [country]);
 
   const onCreateNewRegionVisit = useCallback(() => {
     modal.setIsOpen(true);
@@ -180,8 +78,13 @@ export const CountryPopup = ({ country }) => {
   const countColumnSpace = Math.floor(12 / regionStatistics.filter(stats => stats.show).length);
 
   useEffect(() => {
-    console.log(country)
-    if(!modal.isOpen){
+    if (user?.id) {
+      getHistoryDataForRegion(country.id);
+    }
+  }, [country]);
+
+  useEffect(() => {
+    if(!modal.isOpen && user?.id) {
       fetch()
     }
   }, [modal,country]);
@@ -195,17 +98,19 @@ export const CountryPopup = ({ country }) => {
     <div style={{ color: "white" }}>
       <Col className={'d-flex justify-content-between align-items-center'}>
         <p className={'m-0'} style={{ fontSize: '10px', paddingLeft: '1.2rem' }}>{country.country}</p>
-        <button onClick={onCreateNewRegionVisit} className='btn btn-secondary py-0 px-0' style={{ fontSize: '10px' }}>
-          <span>Add new visit</span>
-          <CompassOutlined className={'ms-2'}/>
-        </button>
+        {user?.id && (
+          <button onClick={onCreateNewRegionVisit} className='btn btn-secondary py-0 px-0' style={{fontSize: '10px'}}>
+            <span>Add new visit</span>
+            <CompassOutlined className={'ms-2'}/>
+          </button>
+        )}
       </Col>
-      <h6 className={'d-flex fw-bold gap-2'} style={{ fontSize: '12px' }}>
-        <FavouriteTag country={country.uname} />
+      <h6 className={'d-flex fw-bold gap-2'} style={{fontSize: '12px'}}>
+        <FavouriteTag country={country.uname}/>
         {country.region}
       </h6>
       <Row className='mt-3'>
-        {regionStatistics.map((statistic, idx) => statistic.show && (
+      {regionStatistics.map((statistic, idx) => statistic.show && (
           <Col key={idx} xs={countColumnSpace} className={`d-flex flex-column m-0 ${idx === 0 ? 'p-0' : 'pe-0 ps-4 py-0'}`}>
             <h6 className={'w-100 fa-xs'}>{statistic.label}</h6>
             <h5 className='text-truncate' style={{ fontSize: '0.875rem' }}>{statistic.text}</h5>
@@ -229,7 +134,7 @@ export const CountryPopup = ({ country }) => {
         ))}
       </div>
 
-      {user && (
+      {user?.id && (
         <>
           <Col style={{ fontSize: '10px' }} className='d-flex justify-content-between align-items-center'>
             <h6 style={{ fontSize: '10px' }}>In groups</h6>
