+++ src/views/Personalnformation/PersonalInformation.jsx
@@ -12,7 +12,8 @@ export const usePersonalInfoModal = create((set) => ({
 
 const PersonalInformation = () => {
   const {isOpen, setIsOpen} = usePersonalInfoModal();
-  const [{data: personalInfo, loading}] = useLoadMe();
+  const [{ data: personalInfo, loading}] = useLoadMe();
+
   if (loading) {
     return null;
   }
